requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
redis
rq==1.16.1
python-jose[cryptography]
bcrypt
python-dotenv
cryptography